+++ r-package/R/detailed_itineraries.R
@@ -219,7 +219,7 @@ detailed_itineraries <- function(r5r_network,
   set_verbose(r5r_network, verbose)
   set_progress(r5r_network, progress)
   set_fare_structure(r5r_network, fare_structure)
-  set_max_fare(r5r_network, max_fare)
+  set_max_fare(r5r_network, max_fare, fare_structure)
   set_output_dir(r5r_network, output_dir)
   set_suboptimal_minutes(
     r5r_network,